% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A \code{data.frame} consisting of two variables. \code{w} represents the object's weight and \code{v} is the value.}

\item{W}{the maximum capacity of the knapsack.}

\item{parallel}{\code{FALSE} by default. If \code{TRUE}, the function is parallelized over detected cores.}
}
\value{
A \code{list} containing the maximum value and the elements.
}
\description{
Return the maximum value given the capacity of knapsack by looking through every possible alternatives.
}
\references{
\url{en.wikipedia.org/wiki/Knapsack_problem}
}
