% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A \code{data.frame} consisting of two variables. \code{w} represents the object's weight and \code{v} is the value.}

\item{W}{the maximum capacity of the knapsack.}
}
\value{
A \code{list} containing the maximum value and the elements.
}
\description{
Returns the maximum value given the capacity of knapsack by sorting the items in decreasing order of value per unit of weight.
}
\references{
\url{en.wikipedia.org/wiki/Knapsack_problem}
}
