% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack Dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A \code{data.frame} consisting of two variables. \code{w} represents the object's weight and \code{v} is the value.}

\item{W}{the maximum capacity of the knapsack.}
}
\value{
A \code{list} containing the maximum value and the elements.
}
\description{
Returns the maximum value given the capacity of knapsack by using dynamic programming.
}
\references{
\url{en.wikipedia.org/wiki/Knapsack_problem}
}
