\name{lab6group8-package}
\alias{lab6group8-package}
\alias{lab6group8}
\docType{package}
\title{
\packageTitle{lab6group8}
}
\description{
\packageDescription{lab6group8}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab6group8}
\packageIndices{lab6group8}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{lab6group8}

Maintainer: \packageMaintainer{lab6group8}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
